Removed some steps from seed script, improved errors, and renamed some functions that have long names

+++ script/scrapeJobs/scrapeJobs.ts
@@ -4,7 +4,7 @@ import {
   buildData,
   hasJobChanged,
 } from "../helperUtilities/helperUtilities.js";
-import { scrapingJobSitesFunction } from "../scrapingJobsSitesFunction/scrapingJobSitesFunction.js";
+import { scrapingJobsFunction } from "../scrapingJobsDetailsFunction/scrapingJobsDetailsFunction.js";
 
 (async () => {
   try {
@@ -16,60 +16,62 @@ import { scrapingJobSitesFunction } from "../scrapingJobsSitesFunction/scrapingJ
       },
     });
 
-    for (const company of companies) {
-      const scrapedJobs = await scrapingJobSitesFunction(company);
+    if (companies.length > 0) {
+      for (const company of companies) {
+        const scrapedJobs = await scrapingJobsFunction(company);
 
-      const existingJobsMap = new Map(
-        company.jobs.map((job) => [job.anchorHref, job]),
-      );
+        const existingJobsMap = new Map(
+          company.jobs.map((job) => [job.anchorHref, job]),
+        );
 
-      const existingJobsIds = new Set(company.jobs.map((job) => job.id));
+        const existingJobsIds = new Set(company.jobs.map((job) => job.id));
 
-      const scrapedJobsIds: Set<number> = new Set();
+        const scrapedJobsIds: Set<number> = new Set();
 
-      await prisma.$transaction(async (tx) => {
-        for (const scrapedJob of scrapedJobs) {
-          const existingJob = existingJobsMap.get(scrapedJob.anchorHref!);
+        await prisma.$transaction(async (tx) => {
+          for (const scrapedJob of scrapedJobs) {
+            const existingJob = existingJobsMap.get(scrapedJob.anchorHref!);
 
-          if (existingJob) {
-            scrapedJobsIds.add(existingJob.id);
+            if (existingJob) {
+              scrapedJobsIds.add(existingJob.id);
 
-            const scrapedJobChanged = hasJobChanged(
-              existingJob!,
-              scrapedJob as Jobs,
-            );
+              const scrapedJobChanged = hasJobChanged(
+                existingJob!,
+                scrapedJob as Jobs,
+              );
 
-            if (scrapedJobChanged) {
-              await tx.jobs.update({
-                where: {
-                  id: existingJob.id,
-                },
+              if (scrapedJobChanged) {
+                await tx.jobs.update({
+                  where: {
+                    id: existingJob.id,
+                  },
+                  data: buildData(scrapedJob as Jobs),
+                });
+              }
+            } else {
+              await tx.jobs.create({
                 data: buildData(scrapedJob as Jobs),
               });
             }
-          } else {
-            await tx.jobs.create({
-              data: buildData(scrapedJob as Jobs),
+          }
+
+          const jobsToDelete = Array.from(
+            existingJobsIds.difference(scrapedJobsIds),
+          );
+
+          // console.log(jobsToDelete);
+
+          if (jobsToDelete.length > 0) {
+            await tx.jobs.deleteMany({
+              where: {
+                id: {
+                  in: jobsToDelete,
+                },
+              },
             });
           }
-        }
-
-        const jobsToDelete = Array.from(
-          existingJobsIds.difference(scrapedJobsIds),
-        );
-
-        // console.log(jobsToDelete);
-
-        if (jobsToDelete.length > 0) {
-          await tx.jobs.deleteMany({
-            where: {
-              id: {
-                in: jobsToDelete,
-              },
-            },
-          });
-        }
-      });
+        });
+      }
     }
   } catch (error) {
     console.error(`Failed to sync company:`, error);
